+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * add `UniqueInSlice` function with generic parameters in `basicalter` package to replace `UniqueStrings` function which is now deprecated
 * add `DelInSlice` function with generic parameters in `basicalter` package to replace `DelStringInSlice` function which is now deprecated
 * add `FilterInSliceWith` function with generic parameters in `basicalter` package to replace `FilterStringsWith` function which is now deprecated
+* add `ReverseSlice` function with generic parameters in `basicalter` package to replace `ReverseStrings` function which is now deprecated
 
 ## v0.5.0